Fixed bugs in this file
This file will occur a Compile Error when I compile it. The wrong code is in Line 19. The full error message is shown below:
checkassert.h: In function ‘bool A_test(std::string, int, std::string)’:
checkassert.h:19:52: error: ‘s’ was not declared in this scope
   19 |    check = check && (r[j] == '0' || r[j] == '1' || s[j] == t[j]);
      |                                                    ^
checkassert.h:19:60: error: ‘t’ was not declared in this scope
   19 |    check = check && (r[j] == '0' || r[j] == '1' || s[j] == t[j]);